+++ doc/changelog.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* :feature:`-` remove deprecated ``PactBase.finished()`` API
 * :feature:`-` write wait's duration seconds to log
 * :feature:`-` Switch to PBR
 * :release:`1.10.0 <15-10-2017>`